Game is up and running using the new modules! * added 'basedir' option in game.ini * added NetComponent for network streaming of entities * removed Actor class * fixed several bugs in the entity and effect modules

+++ src/core/graphics/animator.h
@@ -42,7 +42,7 @@ namespace Graphics
       ~Animator();
 
     // get/set
-      void setAnimationSpeed(int milliseconds);
+      void setAnimationSpeed(float milliseconds);
 
       int  getAnimation() const;
       bool setAnimation(int animation);
